Activation Functions
The Activation Functions is used to calculate the output response of a neuron. The sum of the weighted input signal is applied with an activation to obtain the response. For neurons in same layer, same activation functions are used. There may be linear as well as nonlinear activation functions. The non-linear activation functions are used in a multilayer net. A few linear and non linear activation functions are discussed here,