Create ANDNOT function using Mcculloch-Pitts neuron
McCulloch-Pitts neuron allows binary 0 or 1 states only, i.e it is binary activated. These neurons are connected by direct weighted path. The connected path can be excitatory or inhibitory. Excitatory connections have positive weights and inhibitory connections have negative weights. There will be same weights for the excitatory connection entering into a particular neuron. The neuron is associated with the threshold value. The neuron fires if the net input to the neuron is greater than the threshold. The threshold is set so that the inhibition is absolute, because, non-zero inhibitory input will prevent the neuron from firing. It takes only one time step for a signal to pass over one connection link.

The truth table for the ANDNOT function is, 

                       x1               x2               y
                       0                 0                0
                       0                 1                0
                       1                 0                1
                       1                 1                0

The net input is,

y(in) = x1*w1 + x2*w2
y(in) = x1*1 + x2*(-1)
y(in) = x1 - x2


The output activation is given as, 
y  = f(yin) = 1  if  y(-in) >=1
                 = 0 if  y(-in)  <  1